requirements.txt
PyQt5 >= 5
scapy_python3 >= 0.21